move some common code to a base class we can extend
